requirements.txt
PyYAML>=6.0
fastapi>=0.109.0
uvicorn[standard]>=0.26.0
pydantic>=2.5.3
langchain-core>=0.1.22
langchain-google-vertexai>=0.1.7
google-cloud-aiplatform>=1.38.2
psutil>=2.4.1
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
qdrant-client>=1.9.0
numpy>=1.24.0
python-dotenv